Refactored files to subdirectories
